Refactoring and changing folder structure
