+++ frontend/Dockerfile
@@ -1,5 +1,5 @@
 
-FROM node:8.10 as builder
+FROM node:12.16.2 as builder
 
 COPY . /app
 WORKDIR /app
